requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via python-binance
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.21.0
    # via -r requirements.in
alabaster==1.0.0
    # via sphinx
attrs==25.3.0
    # via aiohttp
babel==2.17.0
    # via sphinx
binance-connector==3.12.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
cryptography==44.0.2
    # via -r requirements.in
dateparser==1.2.1
    # via python-binance
docutils==0.21.2
    # via
    #   sphinx
    #   sphinx-rtd-theme
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-sock
flask-cors==5.0.1
    # via -r requirements.in
flask-sock==0.7.0
    # via -r requirements.in
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   yarl
imagesize==1.4.1
    # via sphinx
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   sphinx
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
numpy==2.2.5
    # via
    #   -r requirements.in
    #   pandas
packaging==25.0
    # via
    #   pytest
    #   sphinx
pandas==2.2.3
    # via
    #   -r requirements.in
    #   pandas-ta
pandas-ta==0.3.14b0
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pycryptodome==3.22.0
    # via
    #   binance-connector
    #   python-binance
pygments==2.19.1
    # via sphinx
pyjwt==2.10.1
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.26.0
    # via -r requirements.in
python-binance==1.0.28
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   dateparser
    #   pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via
    #   dateparser
    #   pandas
regex==2024.11.6
    # via dateparser
requests==2.32.3
    # via
    #   -r requirements.in
    #   binance-connector
    #   python-binance
    #   sphinx
roman-numerals-py==3.1.0
    # via sphinx
simple-websocket==1.1.0
    # via flask-sock
six==1.17.0
    # via
    #   python-binance
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==8.2.3
    # via
    #   -r requirements.in
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-rtd-theme==3.0.2
    # via -r requirements.in
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlalchemy==2.0.40
    # via -r requirements.in
types-requests==2.32.0.20250328
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   aiosqlite
    #   sqlalchemy
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via dateparser
urllib3==2.4.0
    # via
    #   requests
    #   types-requests
websocket-client==1.8.0
    # via binance-connector
websockets==15.0.1
    # via
    #   -r requirements.in
    #   python-binance
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wsproto==1.2.0
    # via simple-websocket
yarl==1.20.0
    # via aiohttp
